+++ test/rethinkdb/core_test.clj
@@ -397,8 +397,10 @@
     (is (thrown? ExceptionInfo (r/run (r/table :nope) conn)))
     (try (r/run (r/table :nope) conn)
          (catch ExceptionInfo ex
-           (let [{:keys [r]} (ex-data ex)
+           (let [r (get-in (ex-data ex) [:response :r])
+                 etype (:type (ex-data ex))
                  msg (.getMessage ex)]
+             (is (= etype :op-failed))
              (is (= r ["Table `cljrethinkdb_test.nope` does not exist."]))
              (is (= "RethinkDB server: Table `cljrethinkdb_test.nope` does not exist." msg)))))))